# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NicheFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /business:
    get:
      operationId: getBusiness
      tags: [business]
      summary: Get business profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Business not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createBusiness
      tags: [business]
      summary: Create business profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessRequest"
      responses:
        "201":
          description: Business created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    put:
      operationId: updateBusiness
      tags: [business]
      summary: Update business profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessRequest"
      responses:
        "200":
          description: Business updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
...
  /public/{slug}:
    get:
      operationId: getPublicPage
      tags: [business]
      summary: Get public business page
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public business page data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPage"
        "404":
          description: Business not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Business:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - category
        - currency
        - createdAt
    CreateBusinessRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        currency:
          type: string
      required:
        - name
        - category
    UpdateBusinessRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        currency:
          type: string
        logoUrl:
          type: string
...
    PublicPage:
      type: object
      properties:
        business:
          $ref: "#/components/schemas/Business"
        services:
          type: array
          items:
            type: string
      required:
        - business
        - services